proposed changes to the plugins folder to cleanup the code
