+++ hpotter-server/hpotter/__main__.py
@@ -1,7 +1,7 @@
 import signal, sys, inspect, os
 
 import hpotter.plugins
-from hpotter.plugins.plugin_handler import start_plugins, stop_plugins
+from hpotter.plugins.handler import start_plugins, stop_plugins
 from hpotter.env import logger, stop_shell, close_db
 
 def shutdown_servers(signum, frame):
